requirements.txt
streamlit>=1.18
pandas>=1.5
folium>=0.14
geopy>=2.3
requests
python-dateutil>=2.8
sqlite3-binary
